requirements.txt
requests
openai
python-dotenv
csv
snowflake-connector-python
schedule

